修复了之前的一些bug

+++ user/views.py
@@ -2,8 +2,10 @@
 
 from django.views import View
 from django.http import JsonResponse, FileResponse
+import logging
 
 import hashlib
+import os
 
 from user.models import User
 from park_backend import settings
@@ -11,12 +13,12 @@ import numpy as np
 
 from utils.respones import CommonResponseMixin, ReturnCode
 from utils import auth
-
-import random
-from django_redis import get_redis_connection
+from utils.parse import image_url, sex_to_int, int_to_sex
 
 DUE = 900
 INTERNAL = 60
+
+logger = logging.getLogger(os.path.abspath(__file__))
 
 
 class AuthorizeView(View, CommonResponseMixin):
@@ -31,7 +33,7 @@ class AuthorizeView(View, CommonResponseMixin):
         :return:
         """
 
-        request.session.delete()
+        # request.session.delete()
 
         phone_number = request.POST.get('phone_number')
         password = request.POST.get('password')
@@ -53,20 +55,17 @@ class AuthorizeView(View, CommonResponseMixin):
             response = AuthorizeView.wrap_json_response(code=ReturnCode.WRONG_PASSWORD)
             return JsonResponse(data=response, safe=False)
         else:
-            request.session['phone_number'] = phone_number
-            request.session['is_authorized'] = True
+            # request.session['phone_number'] = phone_number
             response = {
-                'phone_number': user.password,
+                'phone_number': user.phone_number,
                 'petname': user.nickname,
                 'name': user.name,
                 'ID': user.ID,
-                'sex': '男' if user.sex == 1 else '女',
+                'sex': int_to_sex(user.sex),
                 'bankcard': user.bankcard,
-                'icon_url': user.icon.url if user.icon.url != '/media/' + settings.default_icon
-                else '/' + settings.default_icon,
+                'icon_url': image_url(user.icon.url),
                 'state': user.state,
             }
-
 
             response = AuthorizeView.wrap_json_response(data=response)
             return JsonResponse(data=response, safe=False)
@@ -131,17 +130,29 @@ class RegisterView(View, CommonResponseMixin):
         phone_number = request.POST.get('phone_number')
         password = request.POST.get('password')
         sex = request.POST.get('sex')
+        icon = request.FILES.get('icon')
 
         # if nickname is None or phone_number is None or password is None or sex is None:
         if not all([nickname, phone_number, password, sex]):
             response = RegisterView.wrap_json_response(code=ReturnCode.BROKEN_PARAMS)
             return JsonResponse(data=response, safe=False)
 
-        user = User(phone_number=phone_number, password=password, nickname=nickname, sex=1 if sex == "男" else 0)
+        user = User.objects.filter(phone_number=phone_number).first()
+        if user is not None:
+            response = self.wrap_json_response(code=ReturnCode.PHONE_NUMBER_EXISTED)
+            return JsonResponse(data=response, safe=False)
 
+        user = User.objects.create(phone_number=phone_number, password=password, nickname=nickname, sex=sex_to_int(sex))
+        if icon is not None:
+            user.icon = icon
+            user.save()
         user.save()
 
-        response = RegisterView.wrap_json_response(code=ReturnCode.SUCCESS)
+        data = {
+            'image_url': image_url(user.icon.url)
+        }
+
+        response = RegisterView.wrap_json_response(code=ReturnCode.SUCCESS, data=data)
         return JsonResponse(data=response, safe=False)
 
 
@@ -149,55 +160,14 @@ class UserManageView(View, CommonResponseMixin):
 
     @auth.login_required
     def post(self, request):
-
-        phone_number = request.session['phone_number']
-        new_phone_number = request.POST.get('phone_number')
+        phone_number = request.POST.get('phone_number')
         nickname = request.POST.get('petname')
-        icon = request.FILES.get('icon')
         password = request.POST.get('password')
         sex = request.POST.get('sex')
-
+        icon = request.FILES.get('icon')
         if nickname is None and password is None and sex is None and phone_number is None:
-            response = RegisterView.wrap_json_response(code=ReturnCode.BROKEN_PARAMS)
+            response = self.wrap_json_response(code=ReturnCode.BROKEN_PARAMS)
             return JsonResponse(data=response, safe=False)
-
-        sex = 1 if sex == "男" else 0
-
-        user = User.objects.filter(phone_number=phone_number).first()
-
-        icon_md5 = hashlib.md5(np.array(icon))
-        user_icon_md5 = hashlib.md5(np.array(user.icon))
-        if user_icon_md5 != icon_md5 and user_icon_md5 != settings.default_md5:
-            user.icon.storage.delete(user.icon.path)
-            user.icon = icon
-        elif user_icon_md5 == settings.default_md5:
-            user.icon = icon
-
-        user.password = password if password is not None else user.password
-        user.sex = sex if sex is not None else user.sex
-        user.nickname = nickname if nickname is not None else user.nickname
-        user.phone_number = new_phone_number if phone_number is not None else user.phone_number
-        user.save()
-
-        response = RegisterView.wrap_json_response(code=ReturnCode.SUCCESS)
-        return JsonResponse(data=response, safe=False)
-
-
-class IconView(View):
-
-    @auth.login_required
-    def get(self, request):
-        phone_number = request.session['phone_number']
-        user = User.objects.filter(phone_number=phone_number).first()
-
-        return FileResponse(user.icon)
-
-
-class GetInfoView(View, CommonResponseMixin):
-
-    @auth.login_required
-    def get(self, request):
-        phone_number = request.session.get('phone_number')
 
         try:
             user = User.objects.get(phone_number=phone_number)
@@ -205,20 +175,37 @@ class GetInfoView(View, CommonResponseMixin):
             response = self.wrap_json_response(code=ReturnCode.PHONE_NUMBER_NOT_EXISTED)
             return JsonResponse(data=response, safe=False)
 
-        response = {
-            'phone_number': user.password,
-            'petname': user.nickname,
-            'name': user.name,
-            'ID': user.ID,
-            'sex': '男' if user.sex == 1 else '女',
-            'bankcard': user.bankcard,
-            'icon_url': user.icon.url if user.icon.url != '/media/' + settings.default_icon
-            else '/' + settings.default_icon,
-            'state': user.state,
+        if icon is not None:
+            icon_md5 = hashlib.md5(np.array(icon))
+            user_icon_md5 = hashlib.md5(np.array(user.icon))
+            if user_icon_md5 != icon_md5 and user_icon_md5 != settings.default_md5:
+                user.icon.storage.delete(user.icon.path)
+                user.icon = icon
+            elif user_icon_md5 == settings.default_md5:
+                user.icon = icon
+
+            user.save()
+
+        data = {
+            'image_url': image_url(user.icon.url)
         }
 
-        response = AuthorizeView.wrap_json_response(data=response)
+        user.password = password if password is not None else user.password
+        user.sex = sex_to_int(sex) if sex is not None else user.sex
+        user.nickname = nickname if nickname is not None else user.nickname
+        user.save()
+        response = self.wrap_json_response(code=ReturnCode.SUCCESS, data=data)
         return JsonResponse(data=response, safe=False)
+
+
+class IconView(View):
+
+    @auth.login_required
+    def get(self, request):
+        phone_number = request.GET.get['phone_number']
+        user = User.objects.filter(phone_number=phone_number).first()
+
+        return FileResponse(user.icon)
 
 
 class NewIconView(View, CommonResponseMixin):
@@ -230,43 +217,45 @@ class NewIconView(View, CommonResponseMixin):
         if not all([icon, phone_number]):
             response = self.wrap_json_response(code=ReturnCode.BROKEN_PARAMS)
             return JsonResponse(data=response, safe=False)
-
         try:
             user = User.objects.get(phone_number=phone_number)
         except User.DoesNotExist:
             response = self.wrap_json_response(code=ReturnCode.PHONE_NUMBER_NOT_EXISTED)
             return JsonResponse(data=response, safe=False)
 
-        icon_md5 = hashlib.md5(np.array(icon))
-        user_icon_md5 = hashlib.md5(np.array(user.icon))
-        if user_icon_md5 != icon_md5 and user_icon_md5 != settings.default_md5:
-            user.icon.storage.delete(user.icon.path)
-            user.icon = icon
-        elif user_icon_md5 == settings.default_md5:
-            user.icon = icon
+        if icon is not None:
+            icon_md5 = hashlib.md5(np.array(icon))
+            user_icon_md5 = hashlib.md5(np.array(user.icon))
+            if user_icon_md5 != icon_md5 and user_icon_md5 != settings.default_md5:
+                user.icon.storage.delete(user.icon.path)
+                user.icon = icon
+                user.save()
+            elif user_icon_md5 == settings.default_md5:
+                user.icon = icon
+                user.save()
+        data = {
+            'image_url': image_url(user.icon.url)
+        }
 
-        user.save()
-
-        response = self.wrap_json_response(code=ReturnCode.SUCCESS)
+        response = self.wrap_json_response(code=ReturnCode.SUCCESS, data=data)
         return JsonResponse(data=response, safe=False)
 
-
-class LogoutView(View, CommonResponseMixin):
-
-    @auth.login_required
-    def post(self, request):
-        phone_number = request.POST.get('phone_number')
-
-        if phone_number is None:
-            response = self.wrap_json_response(code=ReturnCode.BROKEN_PARAMS)
-            return JsonResponse(data=response, safe=False)
-
-        if phone_number != request.session.get('phone_number'):
-            response = self.wrap_json_response(code=ReturnCode.PHONE_NUMBER_NOT_THE_SAME)
-            return JsonResponse(data=response)
-
-        del request.session['phone_number']
-        del request.session['is_authorized']
-
-        response = self.wrap_json_response(code=ReturnCode.SUCCESS)
-        return JsonResponse(data=response, safe=False)
+# class LogoutView(View, CommonResponseMixin):
+#
+#     @auth.login_required
+#     def post(self, request):
+#         phone_number = request.POST.get('phone_number')
+#
+#         if phone_number is None:
+#             response = self.wrap_json_response(code=ReturnCode.BROKEN_PARAMS)
+#             return JsonResponse(data=response, safe=False)
+#
+#         # if phone_number != request.session.get('phone_number'):
+#         #     response = self.wrap_json_response(code=ReturnCode.PHONE_NUMBER_NOT_THE_SAME)
+#         #     return JsonResponse(data=response)
+#
+#         # del request.session['phone_number']
+#         # del request.session['is_authorized']
+#
+#         response = self.wrap_json_response(code=ReturnCode.SUCCESS)
+#         return JsonResponse(data=response, safe=False)
